Добавлен более вместительный тип данных

+++ Учебная практика/2/Задача 552/Data/TaskSolution.cs
@@ -6,7 +6,7 @@ namespace Data
     {
         static void Main()
         {
-            int countAnimalsTypes = int.Parse(Console.ReadLine());
+            Int64 countAnimalsTypes = Int64.Parse(Console.ReadLine());
             string[] numbers = Console.ReadLine().Split(' ');
             int[] countAnimals = new int[numbers.Length];
             for (int index = 0; index < numbers.Length; index++)
@@ -14,7 +14,7 @@ namespace Data
                 countAnimals[index] = int.Parse(numbers[index]);
             }
 
-            int[] combinations = new int [3];
+            Int64[] combinations = new Int64 [3];
             if (countAnimalsTypes > 2)
             {
                 for (int index = 0; index < countAnimalsTypes; index++)
